requirements.txt
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
colorama==0.4.6
ffmpeg-python==0.2.0
filelock==3.12.0
future==0.18.3
global-hotkeys==0.0.4
idna==3.4
Jinja2==3.1.2
keyboard==0.13.5
llvmlite==0.40.0
MarkupSafe==2.1.2
more-itertools==9.1.0
mpmath==1.3.0
networkx==3.1
numba==0.57.0
numpy==1.24.3
openai-whisper==20230314
Pillow==9.3.0
pycparser==2.21
pywin32==306
regex==2023.5.5
requests==2.30.0
scipy==1.10.1
sounddevice==0.4.6
sympy==1.12
tiktoken==0.3.1
torch==1.9.1+cu111
torchaudio==0.9.1
torchvision==0.10.1+cu111
tqdm==4.65.0
typing_extensions==4.5.0
urllib3==2.0.2
